FIX: refactoring contacts ctrl code
